projects/neurolens-app/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2
biopython==1.83
cyvcf2==0.30.28
pysam==0.22.0
reportlab==4.0.7
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
requests==2.31.0
werkzeug==3.0.1
